Fix the client build, which could not configure or compile at all
None of this had ever run: CI's only client job is the web one, it is
marked allowed-to-fail, and the repository had never been pushed, so four
separate blockers sat undetected in a build nothing had executed.

- cmake/Shaders.cmake did `include(bgfxToolUtils)`, but bgfx.cmake keeps
  that module in its own cmake/ directory and does not put it on
  CMAKE_MODULE_PATH for FetchContent consumers. Configure failed outright.

- bgfx_compile_shader_to_header emits one header *per profile*
  (vs_chunk.sc.glsl.bin.h and friends), not the single aggregate
  vs_chunk.sc.bin.h that render/renderer.cpp includes. Nothing ever
  created that file. It is now generated, pulling in exactly the profiles
  BGFX_EMBEDDED_SHADER references on the target platform — which is a
  different set from the ones the host toolchain can build. On Linux the
  macro names a DXBC blob that shaderc there cannot produce and that
  Direct3D11 will never select, so it gets an inert placeholder rather
  than a backend we are pretending to have.

- shaderc was invoked with no include path. bgfx.cmake only sets
  BGFX_SHADER_INCLUDE_PATH in its installed package config, so every
  shader failed on "`u_modelViewProj' undeclared".

- vs_chunk.sc named a local `packed`, which is a reserved word in GLSL.

Also fixes a Approx(1.0f) in test_wave_director.cpp that fails the clang
job: the project builds with -Wdouble-promotion and -Werror, and Approx
takes a double, so the literal promoted implicitly.

Shaders now compile for GLSL, ESSL and SPIR-V, and the client links.

+++ tests/test_wave_director.cpp
@@ -122,7 +122,7 @@ TEST_CASE("late waves get harder through elites, not through more bodies") {
     WaveDirector director(13);
 
     // Early waves are well under the cap, so no elite scaling is needed.
-    CHECK(director.spec_for(3).elite_multiplier == doctest::Approx(1.0f));
+    CHECK(static_cast<f64>(director.spec_for(3).elite_multiplier) == doctest::Approx(1.0));
 
     // Deep waves saturate the cap and must escalate by power instead.
     const WaveSpec deep = director.spec_for(50);
